src/data_pipeline/pyproject.toml
[project]
name = "data-pipeline"
version = "0.1.0"
description = "Data pipeline for Northwind Commerce knowledge assistant"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "sqlite-vec>=0.1.0",
    "nltk>=3.8.0",
    "mistune>=3.0.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.0.14",
    "sqlmodel>=0.0.14",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "ty>=0.0.17",
    "pre-commit>=3.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[project.scripts]
data-pipeline = "main:cli"

[tool.setuptools]
packages = ["database", "processors", "services", "utils"]
py-modules = ["main", "config"]

# So ty finds first-party modules when run from this dir
[tool.ty.environment]
root = ["."]

# Editable-install layout: ty often can't resolve in-tree packages (database, processors, services).
# Runtime and pytest resolve them; ignore here so pre-commit/check-all passes.
[tool.ty.rules]
unresolved-import = "ignore"
unresolved-attribute = "ignore"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
